requirements.txt
alembic==1.2.0
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
get==2019.4.13
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
Pillow==6.1.0
post==2019.4.13
psycopg2==2.8.3
public==2019.4.13
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-http-client==3.2.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
sendgrid==6.1.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.5
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
